# PERF: MultiIndex.get_loc