- [x] closes #29311


```
In [4]: mi_med = pd.MultiIndex.from_product( 
    ...:     [np.arange(1000), np.arange(10), list("A")], names=["one", "two", "three"] 
    ...: ) 

In [5]: %timeit mi_med.get_loc((999, 9, "A"))   
master --> 42.6 µs ± 411 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
PR     --> 11.4 µs ± 120 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```